refactor index.ts by moving seeding invocations into seed functions
